requirements.txt
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
openai==1.3.7
pandas==2.1.3
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
anyio==3.7.1
azure-storage-blob==12.19.0
openpyxl==3.1.2
jinja2==3.1.2 